pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "gnews>=0.4.2",
    "python-dateutil>=2.9.0.post0",
    "requests>=2.32.4",
]
